Extend packed iterator tests

+++ core/utils/math_utils.hpp
@@ -22,6 +22,7 @@
 
 #include <numeric>
 #include <cassert>
+#include <cmath>
 
 NS_ROOT
 NS_BEGIN( math )
@@ -31,7 +32,7 @@ inline uint64_t ceil64(uint64_t value, uint64_t step) {
 }
 
 inline uint32_t ceil32(uint32_t value, uint32_t step) {
-  return uint32_t(std::ceilf(float_t(value)/step))*step;
+  return uint32_t(std::ceil(float_t(value)/step))*step;
 }
 
 IRESEARCH_API uint32_t log2_64( uint64_t value );
